Improvement to add repository dialog
The message field it is now shown only when there is an error with the
URL. Also it was made like a text area instead of single line.
The dialog title was updated with a more according name.
The dialog was resized a little bit as it was too big.